+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 * now all user data is stored in the database
 * psql credentials are now written out via Vault Database Engine
 #### 🚀 Features
+* bump workflow version to `1.2.9`
 * bump vault-package to major version `3.0.0`
 * bump users-package to major version `3.0.0`
 * bump telegram-package to major version `2.0.1`