+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
-## v2.3.0 - 2024-09-11
+## v2.3.0 - 2024-09-12
 ### What's Changed
 **Full Changelog**: https://github.com/obervinov/pyinstabot-downloader/compare/v2.2.1...v2.3.0 by @obervinov in https://github.com/obervinov/pyinstabot-downloader/pull/95
 #### 💥 Breaking Changes
@@ -12,7 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 #### 🚀 Features
 * bump workflow version to `1.2.9`
 * bump vault-package to major version `3.0.0`
-* bump users-package to major version `3.0.2`
+* bump users-package to major version `3.0.3`
 * bump telegram-package to major version `2.0.1`
 * add tests for database and metrics modules
 * [Switch reading of the database connection configuration to db engine](https://github.com/obervinov/pyinstabot-downloader/issues/33)